requirements.txt
opencv-python
pandas
ultralytics
matplotlib
Pillow
numpy
pytesseract
easyocr
openpyxl
